"add improvement after mentor review"

+++ src/main/java/com/example/bookstore/service/cartitem/CartItemServiceImpl.java
@@ -9,6 +9,7 @@ import com.example.bookstore.model.ShoppingCart;
 import com.example.bookstore.repository.book.BookRepository;
 import com.example.bookstore.repository.cartitem.CartItemRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -40,9 +41,14 @@ public class CartItemServiceImpl implements CartItemService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long userId) {
         if (!cartItemRepository.existsById(id)) {
             throw new EntityNotFoundException("Can't delete cartItem by id: " + id);
+        }
+        CartItem cartItem = findById(id);
+        if (!cartItem.getShoppingCart().getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("User does not have permission to "
+                    + "delete this cart item");
         }
         cartItemRepository.deleteById(id);
     }
